+++ src/units-format/force-format.ts
@@ -1,4 +1,4 @@
 import { UnitFormat } from "uom";
 
 export const Newton = UnitFormat.createUnitFormat("N", 2);
-export const PoundForce = UnitFormat.createUnitFormat("lb", 2);
+export const PoundForce = UnitFormat.createUnitFormat("lbf", 2);