+++ src/units-format/torque-format.ts
@@ -1,0 +1,4 @@
+import { UnitFormat } from "uom";
+
+export const NewtonMeter = UnitFormat.createUnitFormat("Nm", 0);
+export const PoundForceFoot = UnitFormat.createUnitFormat("lbf-ft", 0);