+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/dividab/uom/compare/v2.5.1...master)
 
+## [v4.0.0](https://github.com/dividab/uom/compare/v4.0.0...v3.0.0) 2026-01-21
+
+### Changed
+
+- Moved NewtonMeter from Energy to Torque quantity.
+
+### Added
+
+- New Torque unit PoundForceFoot
+
 ## [v3.0.0](https://github.com/dividab/uom/compare/v3.0.0...v2.6.0) 2025-10-24
 
 ### Changed